requirements.txt
Django==1.11.13
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.7.4
PyJWT==1.6.4
pytz==2018.4
